a2a-examples/k8s_a2a_mcp/pyproject.toml
[project]
name = "k8s-a2a-mcp"
version = "0.1.0"
description = "Automated Kubernetes Pod CrashLoopBackOff Diagnosis using A2A and MCP"
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "a2a-sdk>=0.2.9",
    "click>=8.1.8",
    "fastmcp>=1.0",
    "google-adk>=1.0.0",
    "google-cloud-aiplatform>=1.91.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.0.10",
    "langchain-mcp-adapters>=0.0.9",
    "langgraph>=0.4.1",
    "mcp[cli]>=1.5.0",
    "nest-asyncio>=1.6.0",
    "networkx>=3.4.2",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
]

[project.scripts]
k8s-a2a-mcp = "k8s_debug:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/k8s_debug"]